# tests/golden/contracts/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Greenhouse API",
  "description": "Smart plant irrigation system with evidence-based plant care, multi-sensor conflict resolution, and self-learning irrigation profiles.\n\n",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/sensors": {
   "get": {
    "tags": [
     "sensors"
    ],
    "summary": "List sensors across all clusters",
    "description": "List every sensor across all clusters with optional cluster filter and cursor pagination.\n\nArgs:\n    cluster_id: Restrict to a single cluster.\n    limit: Page size (default 100, max 500).\n    cursor: Id-based cursor — pass the previous response's ``next_cursor``\n        to fetch the next page.\n\nReturns:\n    The page of sensors plus a ``next_cursor`` (None when the page was not\n    full and there are no more rows to fetch).",
    "operationId": "list_all_sensors",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Restrict results to a specific cluster",
       "title": "Cluster Id"
      },
      "description": "Restrict results to a specific cluster"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Id cursor — return rows with id > cursor",
       "title": "Cursor"
      },
      "description": "Id cursor — return rows with id > cursor"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SensorListResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/clusters/{cluster_id}/sensors": {
   "post": {
    "tags": [
     "sensors"
    ],
    "summary": "Add Sensor",
    "description": "Register a Tuya sensor under a cluster.\n\nA sensor may optionally be linked to a specific plant; otherwise it is\ntreated as a cluster-level reading.\n\nArgs:\n    cluster_id: Cluster the sensor belongs to.\n    request: Tuya device ID, sensor name, type (the device model key,\n        e.g. `tuya.tr301z`),\n        optional config dict, and optional plant_id for per-plant linking.\n\nRaises:\n    HTTPException: 404 if the cluster (or referenced plant) is missing,\n        409 if the Tuya device ID is already registered.\n\nReturns:\n    The created sensor record.",
    "operationId": "add_sensor",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreateSensorRequest"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SensorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "sensors"
    ],
    "summary": "List Sensors",
    "description": "List every sensor registered to a cluster.\n\nArgs:\n    cluster_id: ID of the cluster to enumerate.\n\nReturns:\n    The cluster's sensors (empty for an unknown cluster).",
    "operationId": "list_sensors",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/SensorResponse"
         },
         "title": "Response List Sensors"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/clusters/{cluster_id}/sensors/{sensor_id}": {
   "get": {
    "tags": [
     "sensors"
    ],
    "summary": "Get a sensor by ID",
    "description": "Fetch a single sensor by ID.\n\nArgs:\n    cluster_id: Cluster the sensor belongs to.\n    sensor_id: Numeric sensor identifier.\n\nReturns:\n    The sensor record.\n\nRaises:\n    HTTPException: 404 if the sensor does not exist or belongs to a\n        different cluster.",
    "operationId": "get_sensor",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     },
     {
      "name": "sensor_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Sensor Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SensorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "sensors"
    ],
    "summary": "Update a sensor",
    "description": "Partially update a sensor's metadata.\n\nOnly fields present in the request body are modified; omitted fields are\nleft unchanged. The sensor must belong to the specified cluster.\n\nArgs:\n    cluster_id: Cluster the sensor belongs to.\n    sensor_id: Numeric sensor identifier.\n    request: Fields to update — any subset of name, type, config,\n        and plant_id.\n\nReturns:\n    The updated sensor.\n\nRaises:\n    HTTPException: 404 if the sensor does not exist or belongs to a\n        different cluster, or if ``plant_id`` is not a plant of that cluster.",
    "operationId": "update_sensor",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     },
     {
      "name": "sensor_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Sensor Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UpdateSensorRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SensorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "sensors"
    ],
    "summary": "Delete a sensor",
    "description": "Delete a sensor and all its historical readings.\n\nThis operation is irreversible. The sensor must belong to the specified\ncluster.\n\nArgs:\n    cluster_id: Cluster the sensor belongs to.\n    sensor_id: Numeric sensor identifier.\n\nReturns:\n    success=True on successful deletion.\n\nRaises:\n    HTTPException: 404 if the sensor does not exist or belongs to a\n        different cluster.",
    "operationId": "delete_sensor",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     },
     {
      "name": "sensor_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Sensor Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SuccessResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sensors/{sensor_id}/assignments": {
   "get": {
    "tags": [
     "sensors"
    ],
    "summary": "List the sensor's plant-assignment history",
    "description": "Return every plant this sensor has ever been linked to, oldest first.\n\nEach row covers the interval ``[started_at, ended_at)``. ``ended_at=None``\non the latest row means the sensor is currently assigned. Used by the UI\nto explain why a plant's history changes shape when a sensor is moved,\nand by audit tooling to verify attribution.\n\nArgs:\n    sensor_id: Numeric sensor identifier.\n\nReturns:\n    Full assignment list (possibly empty if the sensor was never linked).\n\nRaises:\n    HTTPException: 404 if the sensor does not exist.",
    "operationId": "list_sensor_assignments",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "sensor_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Sensor Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SensorAssignmentListResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/clusters/{cluster_id}/history": {
   "get": {
    "tags": [
     "operations"
    ],
    "summary": "History",
    "description": "Return recent sensor readings and irrigation events for a cluster.\n\nArgs:\n    cluster_id: Cluster to inspect.\n    hours: Look-back window in hours (default 24).\n    limit: Maximum readings/events per sensor or irrigator (default 50).\n\nReturns:\n    Per-sensor readings and per-irrigator events within the window.\n\nRaises:\n    HTTPException: 404 if the cluster does not exist.",
    "operationId": "history",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     },
     {
      "name": "hours",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 24,
       "title": "Hours"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HistoryResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CreateSensorRequest": {
    "properties": {
     "tuya_device_id": {
      "type": "string",
      "title": "Tuya Device Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "type": {
      "type": "string",
      "title": "Type"
     },
     "config": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Config"
     },
     "plant_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Plant Id"
     }
    },
    "type": "object",
    "required": [
     "tuya_device_id",
     "name",
     "type"
    ],
    "title": "CreateSensorRequest",
    "description": "Body for registering a sensor in a cluster."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "HistoryResponse": {
    "properties": {
     "cluster_name": {
      "type": "string",
      "title": "Cluster Name"
     },
     "sensors": {
      "items": {
       "$ref": "#/components/schemas/SensorHistoryResponse"
      },
      "type": "array",
      "title": "Sensors"
     },
     "irrigators": {
      "items": {
       "$ref": "#/components/schemas/IrrigatorHistoryResponse"
      },
      "type": "array",
      "title": "Irrigators"
     }
    },
    "type": "object",
    "required": [
     "cluster_name",
     "sensors",
     "irrigators"
    ],
    "title": "HistoryResponse",
    "description": "A cluster's sensor readings and irrigation events over the requested window."
   },
...
   "IrrigationEventResponse": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "irrigator_id": {
      "type": "integer",
      "title": "Irrigator Id"
     },
     "timestamp": {
      "type": "integer",
      "title": "Timestamp"
     },
     "action": {
      "type": "string",
      "title": "Action"
     },
     "duration_minutes": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Duration Minutes"
     },
     "triggered_by": {
      "type": "string",
      "title": "Triggered By"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "id",
     "irrigator_id",
     "timestamp",
     "action",
     "duration_minutes",
     "triggered_by",
     "notes"
    ],
    "title": "IrrigationEventResponse",
    "description": "One irrigation event (start, stop, manual log, …) with what triggered it."
   },
...
   "IrrigatorHistoryResponse": {
    "properties": {
     "irrigator_id": {
      "type": "integer",
      "title": "Irrigator Id"
     },
     "irrigator_name": {
      "type": "string",
      "title": "Irrigator Name"
     },
     "events": {
      "items": {
       "$ref": "#/components/schemas/IrrigationEventResponse"
      },
      "type": "array",
      "title": "Events"
     }
    },
    "type": "object",
    "required": [
     "irrigator_id",
     "irrigator_name",
     "events"
    ],
    "title": "IrrigatorHistoryResponse",
    "description": "One irrigator's events over the requested window."
   },
...
   "SensorAssignmentListResponse": {
    "properties": {
     "sensor_id": {
      "type": "integer",
      "title": "Sensor Id"
     },
     "assignments": {
      "items": {
       "$ref": "#/components/schemas/SensorAssignmentResponse"
      },
      "type": "array",
      "title": "Assignments"
     }
    },
    "type": "object",
    "required": [
     "sensor_id",
     "assignments"
    ],
    "title": "SensorAssignmentListResponse",
    "description": "Full assignment history for a sensor, oldest first."
   },
   "SensorAssignmentResponse": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "sensor_id": {
      "type": "integer",
      "title": "Sensor Id"
     },
     "plant_id": {
      "type": "integer",
      "title": "Plant Id"
     },
     "started_at": {
      "type": "integer",
      "title": "Started At"
     },
     "ended_at": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ended At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "sensor_id",
     "plant_id",
     "started_at",
     "ended_at"
    ],
    "title": "SensorAssignmentResponse",
    "description": "One row from a sensor's plant-assignment history.\n\n``ended_at=None`` means the row is currently active."
   },
   "SensorHistoryResponse": {
    "properties": {
     "sensor_id": {
      "type": "integer",
      "title": "Sensor Id"
     },
     "sensor_name": {
      "type": "string",
      "title": "Sensor Name"
     },
     "readings": {
      "items": {
       "$ref": "#/components/schemas/SensorReadingResponse"
      },
      "type": "array",
      "title": "Readings"
     }
    },
    "type": "object",
    "required": [
     "sensor_id",
     "sensor_name",
     "readings"
    ],
    "title": "SensorHistoryResponse",
    "description": "One sensor's readings over the requested window."
   },
   "SensorListResponse": {
    "properties": {
     "sensors": {
      "items": {
       "$ref": "#/components/schemas/SensorResponse"
      },
      "type": "array",
      "title": "Sensors"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     }
    },
    "type": "object",
    "required": [
     "sensors"
    ],
    "title": "SensorListResponse",
    "description": "Paginated top-level sensor listing across all clusters."
   },
   "SensorReadingResponse": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "sensor_id": {
      "type": "integer",
      "title": "Sensor Id"
     },
     "timestamp": {
      "type": "integer",
      "title": "Timestamp"
     },
     "temperature": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Temperature"
     },
     "soil_moisture": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Soil Moisture"
     },
     "light": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Light"
     },
     "env_humidity": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Env Humidity"
     },
     "battery_state": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Battery State"
     },
     "water_warning": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Water Warning"
     }
    },
    "type": "object",
    "required": [
     "id",
     "sensor_id",
     "timestamp",
     "temperature",
     "soil_moisture",
     "light",
     "env_humidity",
     "battery_state",
     "water_warning"
    ],
    "title": "SensorReadingResponse",
    "description": "One raw sensor reading as persisted (unset metrics are null)."
   },
   "SensorResponse": {
    "properties": {
     "tuya_device_id": {
      "type": "string",
      "title": "Tuya Device Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "type": {
      "type": "string",
      "title": "Type"
     },
     "config": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Config"
     },
     "plant_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Plant Id"
     },
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "cluster_id": {
      "type": "integer",
      "title": "Cluster Id"
     }
    },
    "type": "object",
    "required": [
     "tuya_device_id",
     "name",
     "type",
     "id",
     "cluster_id"
    ],
    "title": "SensorResponse",
    "description": "A sensor as stored, with its cluster and current plant assignment."
   },
...
   "SuccessResponse": {
    "properties": {
     "success": {
      "type": "boolean",
      "title": "Success"
     }
    },
    "type": "object",
    "required": [
     "success"
    ],
    "title": "SuccessResponse",
    "description": "Generic success acknowledgement."
   },
...
   "UpdateSensorRequest": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Type"
     },
     "config": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Config"
     },
     "plant_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Plant Id"
     }
    },
    "type": "object",
    "title": "UpdateSensorRequest",
    "description": "Partial update of a sensor; omitted fields keep their value."
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 },
...
}